pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "librosa>=0.10.2.post1",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "soundfile>=0.13.1",
    "speechrecognition>=3.14.1",
    "streamlit>=1.42.0",
    "trafilatura>=2.0.0",
    "whisper>=1.1.10",
]
